// prisma/generated/type-graphql/resolvers/inputs/Import_export_logWhereInput.ts
export class Import_export_logWhereInput {
  @TypeGraphQL.Field(_type => [Import_export_logWhereInput], {
    nullable: true
  })
  AND?: Import_export_logWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => [Import_export_logWhereInput], {
    nullable: true
  })
  OR?: Import_export_logWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => [Import_export_logWhereInput], {
    nullable: true
  })
  NOT?: Import_export_logWhereInput[] | undefined;

  @TypeGraphQL.Field(_type => BytesFilter, {
    nullable: true
  })
  id?: BytesFilter | undefined;

  @TypeGraphQL.Field(_type => StringFilter, {
    nullable: true
  })
  activity?: StringFilter | undefined;

  @TypeGraphQL.Field(_type => StringFilter, {
    nullable: true
  })
  state?: StringFilter | undefined;

  @TypeGraphQL.Field(_type => IntFilter, {
    nullable: true
  })
  records?: IntFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  username?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  profile_name?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => BytesNullableFilter, {
    nullable: true
  })
  user_id?: BytesNullableFilter | undefined;

  @TypeGraphQL.Field(_type => BytesNullableFilter, {
    nullable: true
  })
  profile_id?: BytesNullableFilter | undefined;

  @TypeGraphQL.Field(_type => BytesNullableFilter, {
    nullable: true
  })
  file_id?: BytesNullableFilter | undefined;

  @TypeGraphQL.Field(_type => DateTimeFilter, {
    nullable: true
  })
  created_at?: DateTimeFilter | undefined;

  @TypeGraphQL.Field(_type => DateTimeNullableFilter, {
    nullable: true
  })
  updated_at?: DateTimeNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  config?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => StringNullableFilter, {
    nullable: true
  })
  result?: StringNullableFilter | undefined;

  @TypeGraphQL.Field(_type => BytesNullableFilter, {
    nullable: true
  })
  invalid_records_log_id?: BytesNullableFilter | undefined;

  @TypeGraphQL.Field(_type => Import_export_fileNullableRelationFilter, {
    nullable: true
  })
  import_export_file?: Import_export_fileNullableRelationFilter | undefined;

  @TypeGraphQL.Field(_type => Import_export_logNullableRelationFilter, {
    nullable: true
  })
  import_export_log?: Import_export_logNullableRelationFilter | undefined;

  @TypeGraphQL.Field(_type => Import_export_logListRelationFilter, {
    nullable: true
  })
  other_import_export_log?: Import_export_logListRelationFilter | undefined;

  @TypeGraphQL.Field(_type => Import_export_profileNullableRelationFilter, {
    nullable: true
  })
  import_export_profile?: Import_export_profileNullableRelationFilter | undefined;

  @TypeGraphQL.Field(_type => UserNullableRelationFilter, {
    nullable: true
  })
  user?: UserNullableRelationFilter | undefined;
}